Mac platform simplified + fixes in EventDispatcher

+++ cocos/2d/CCEventDispatcher.cpp
@@ -234,7 +234,7 @@ void EventDispatcher::visitTarget(Node* node, bool isRootNode)
     
     if (isRootNode)
     {
-        std::vector<int> globalZOrders;
+        std::vector<float> globalZOrders;
         globalZOrders.reserve(_globalZOrderNodeMap.size());
         
         for (const auto& e : _globalZOrderNodeMap)
@@ -242,7 +242,7 @@ void EventDispatcher::visitTarget(Node* node, bool isRootNode)
             globalZOrders.push_back(e.first);
         }
         
-        std::sort(globalZOrders.begin(), globalZOrders.end(), [](const int a, const int b){
+        std::sort(globalZOrders.begin(), globalZOrders.end(), [](const float a, const float b){
             return a < b;
         });
